requirements.txt
celery
djangorestframework
python-dotenv
redis

pdf2image
opencv-python
langchain
docxtpl
pytesseract
openai